fix: freeAmountの計算でオーバーフローが発生する問題を修正

+++ lerna-util-sequence/src/main/scala/lerna/util/sequence/SequenceStore.scala
@@ -163,7 +163,7 @@ private[sequence] object SequenceStore extends AppTypedActorLogging {
   }
   final case class ReserveSequence(
       maxReservedValue: BigInt,
-      reservationAmount: Int,
+      reservationAmount: BigInt,
       sequenceSubId: Option[String],
       replyTo: ActorRef[ReservationResponse],
   ) extends Command {
